Refactor nodegroup creation
